use github app authentication for org.kohsuke.github library calls using GitHubFacade bean
