+++ server/application-server/src/main/resources/application-dev.yml
@@ -37,11 +37,14 @@ github:
     clientId: ${GITHUB_CLIENT_ID:}
     installationId: ${GITHUB_INSTALLATION_ID:}
     privateKeyPath: ${GITHUB_PRIVATE_KEY_PATH:}
+
+http:
     cache:
         enabled: true
         ttl: 500
         # in MB
         size: 50
+
 
 monitoring:
     # List of repositories to monitor in the format owner/repository
